+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Fix item image URL on JSON source.
+
 ## 0.9.1 - 2020-12-03
 
 ### Fixed